Add ControllerRecorder class.
