+++ CharacterController.cpp
@@ -3,386 +3,11 @@
 #include "Character.h"
 #include "Camera.h"
 #include "Control.h"
+#include "Brain.h"
+#include "ControllerRecorder.h"
 #include "Define.h"
 #include "DxLib.h"
 #include <algorithm>
-
-
-// Brainクラス
-Brain::Brain() {
-	m_characterAction_p = NULL;
-}
-
-
-/*
-* キーボード
-*/
-KeyboardBrain::KeyboardBrain(const Camera* camera) {
-	m_camera_p = camera;
-}
-
-void KeyboardBrain::bulletTargetPoint(int& x, int& y) {
-	// マウスの位置
-	int mouseX, mouseY;
-	GetMousePoint(&mouseX, &mouseY);
-
-	// カメラで座標を補正
-	m_camera_p->getMouse(&mouseX, &mouseY);
-
-	x = mouseX;
-	y = mouseY;
-}
-
-// 話しかけたり扉入ったり
-bool KeyboardBrain::actionOrder() {
-	return controlW() == 1;
-}
-
-// 移動（上下左右の入力）
-void KeyboardBrain::moveOrder(int& right, int& left, int& up, int& down) {
-	right = controlD();
-	left = controlA();
-	up = controlW();
-	down = controlS();
-}
-
-// ジャンプ
-int KeyboardBrain::jumpOrder() {
-	return controlSpace();
-}
-
-// しゃがみ
-int KeyboardBrain::squatOrder() {
-	return controlS();
-}
-
-// 近距離攻撃
-int KeyboardBrain::slashOrder() {
-	return rightClick();
-}
-
-// 遠距離攻撃
-int KeyboardBrain::bulletOrder() {
-	return leftClick();
-}
-
-
-/*
-* Normal AI
-*/
-NormalAI::NormalAI() {
-	m_target_p = NULL;
-	m_gx = 0;
-	m_gy = 0;
-	m_rightKey = 0;
-	m_leftKey = 0;
-	m_upKey = 0;
-	m_downKey = 0;
-	m_jumpCnt = 0;
-	m_squatCnt = 0;
-	m_moveCnt = 0;
-}
-
-
-Brain* NormalAI::createCopy(std::vector<Character*> characters, const Camera* camera) {
-	NormalAI* res = new NormalAI();
-
-	if (m_target_p != NULL) {
-		for (unsigned int i = 0; i < characters.size(); i++) {
-			if (m_target_p->getId() == characters[i]->getId()) {
-				res->setTarget(characters[i]);
-				break;
-			}
-		}
-	}
-	setParam(res);
-	return res;
-}
-
-void NormalAI::setParam(NormalAI* brain) {
-	// 移動用
-	brain->setRightKey(m_rightKey);
-	brain->setLeftKey(m_leftKey);
-	brain->setUpKey(m_upKey);
-	brain->setDownKey(m_downKey);
-	// ジャンプの長さ
-	brain->setJumpCnt(m_jumpCnt);
-	// しゃがむ長さ
-	brain->setSquatCnt(m_squatCnt);
-	// 移動目標
-	brain->setGx(m_gx);
-	brain->setGy(m_gy);
-	brain->setMoveCnt(m_moveCnt);
-}
-
-void NormalAI::setCharacterAction(const CharacterAction* characterAction) { 
-	m_characterAction_p = characterAction;
-	// 目標地点は現在地に設定
-	m_gx = m_characterAction_p->getCharacter()->getX();
-	m_gy = m_characterAction_p->getCharacter()->getY();
-}
-
-void NormalAI::bulletTargetPoint(int& x, int& y) {
-	if (m_target_p == NULL) {
-		x = 0;
-		y = 0;
-	}
-	else { // ターゲットに向かって射撃攻撃
-		x = m_target_p->getCenterX() + (GetRand(BULLET_ERROR) - BULLET_ERROR / 2);
-		y = m_target_p->getCenterY() + (GetRand(BULLET_ERROR) - BULLET_ERROR / 2);
-	}
-}
-
-void NormalAI::moveOrder(int& right, int& left, int& up, int& down) {
-	// 現在地
-	int x = m_characterAction_p->getCharacter()->getX();
-	int y = m_characterAction_p->getCharacter()->getY();
-
-	// (壁につっかえるなどで)移動できてないから諦める
-	//DrawFormatString(800, 50, GetColor(255, 255, 255), "moveCnt = %d, x(%d) -> gx(%d)", m_moveCnt, x, m_gx);
-	if (m_moveCnt >= GIVE_UP_MOVE_CNT) {
-		m_gx = x;
-		m_gy = y;
-	}
-
-	// 目標地点設定
-	bool alreadyGoal = m_gx > x - GX_ERROR && m_gx < x + GX_ERROR;
-	// ダメージを受けたらリセット
-	if (m_characterAction_p->getState() == CHARACTER_STATE::DAMAGE) {
-		m_gx = x, m_gy = y;
-	}
-	else if (alreadyGoal && GetRand(MOVE_RAND) == 0) {
-		if (m_target_p != NULL && abs(x - m_target_p->getCenterX()) < TARGET_DISTANCE) {
-			// targetについていく
-			m_gx = m_target_p->getCenterX() + GetRand(2000) - 1000;
-		}
-		else {
-			// ランダムに設定
-			m_gx = GetRand(200) + 100;
-			if (GetRand(99) < GX_ERROR) { m_gx *= -1; }
-			m_gx += x;
-		}
-	}
-	stickOrder(right, left, up, down);
-}
-
-// スティック操作
-void NormalAI::stickOrder(int& right, int& left, int& up, int& down) {
-	// 現在地
-	int x = m_characterAction_p->getCharacter()->getX();
-	int y = m_characterAction_p->getCharacter()->getY();
-
-	// 目標に向かって走る
-	if (m_gx > x + GX_ERROR) {
-		m_rightKey++;
-		m_leftKey = 0;
-		m_moveCnt++;
-	}
-	else if (m_gx < x - GX_ERROR) {
-		m_rightKey = 0;
-		m_leftKey++;
-		m_moveCnt++;
-	}
-	else {
-		m_rightKey = 0;
-		m_leftKey = 0;
-		m_moveCnt = 0;
-	}
-
-	// 反映
-	right = m_rightKey;
-	left = m_leftKey;
-	up = m_upKey;
-	down = m_downKey;
-}
-
-int NormalAI::jumpOrder() {
-	// ダメージを食らったらリセット
-	if (m_characterAction_p->getState() == CHARACTER_STATE::DAMAGE) {
-		m_jumpCnt = 0;
-		if (GetRand(120) == 0) { return 1; }
-	}
-
-	// ランダムでジャンプ
-	if (m_squatCnt == 0 && GetRand(99) == 0) { m_jumpCnt = GetRand(15) + 5; }
-
-	// 壁にぶつかったからジャンプ
-	if (m_rightKey > 0 && m_characterAction_p->getRightLock()) { m_jumpCnt = 20; }
-	else if (m_leftKey > 0 && m_characterAction_p->getLeftLock()) { m_jumpCnt = 20; }
-
-	if (m_jumpCnt > 0) { m_jumpCnt--; }
-	return m_jumpCnt == 0 ? 0 : 20 - m_jumpCnt;
-}
-
-int NormalAI::squatOrder() {
-	// ダメージを食らったらリセット
-	if (m_characterAction_p->getState() == CHARACTER_STATE::DAMAGE) {
-		m_squatCnt = 0;
-	}
-
-	// 目標地点にいないならしゃがまない
-	int x = m_characterAction_p->getCharacter()->getX();
-	bool alreadyGoal = m_gx > x - GX_ERROR && m_gx < x + GX_ERROR;
-	if (!alreadyGoal) { m_squatCnt = 0; }
-
-	// ランダムでしゃがむ
-	if (alreadyGoal && m_characterAction_p->getGrand() && GetRand(99) == 0) { 
-		m_squatCnt = GetRand(120) + 60;
-	}
-
-	if (m_squatCnt > 0) { m_squatCnt--; }
-	return m_squatCnt;
-}
-
-int NormalAI::slashOrder() {
-	if (m_target_p == NULL || m_target_p->getHp() == 0) {
-		return 0;
-	}
-	// 遠距離の敵には斬撃しない
-	if (m_target_p != NULL && abs(m_target_p->getCenterX() - m_characterAction_p->getCharacter()->getCenterX()) > 500) {
-		return 0;
-	}
-	// ランダムで斬撃
-	if (GetRand(50) == 0) {
-		return 1;
-	}
-	return 0;
-}
-
-int NormalAI::bulletOrder() {
-	if (m_target_p == NULL || m_target_p->getHp() == 0) { 
-		return 0;
-	}
-	int x = m_characterAction_p->getCharacter()->getX();
-	if (abs(x - m_target_p->getCenterX()) > TARGET_DISTANCE) {
-		return 0;
-	}
-	// ランダムで射撃
-	if (GetRand(30) == 0) { 
-		return 1;
-	}
-	return 0;
-}
-
-// 攻撃対象を決める(targetのままか、characterに変更するか)
-void NormalAI::searchTarget(const Character* character) {
-	if (GetRand(99) < 50) {
-		int x = m_characterAction_p->getCharacter()->getX();
-		// 距離が遠い
-		if (abs(x - character->getCenterX()) > TARGET_DISTANCE) {
-			return;
-		}
-		// 味方じゃなければ&中立じゃなければ
-		int groupId = character->getGroupId();
-		if (groupId != m_characterAction_p->getCharacter()->getGroupId() && groupId != -1) {
-			m_target_p = character;
-		}
-	}
-}
-
-// 攻撃対象を変更する必要があるならtrueでアピールする。
-bool NormalAI::needSearchTarget() const {
-	if (m_target_p == NULL || m_target_p->getHp() == 0 || GetRand(99) == 0 || m_target_p->getGroupId() == -1) {
-		return true;
-	}
-	return false;
-}
-
-int  NormalAI::getTargetId() const { return m_target_p == NULL ? NULL : m_target_p->getId(); }
-
-
-void ParabolaAI::bulletTargetPoint(int& x, int& y) {
-	if (m_target_p == NULL) {
-		x = 0;
-		y = 0;
-	}
-	else { // ターゲットに向かって射撃攻撃
-		x = m_target_p->getCenterX() + (GetRand(BULLET_ERROR) - BULLET_ERROR / 2);
-		y = m_target_p->getCenterY() + (GetRand(BULLET_ERROR) - BULLET_ERROR / 2) - 200;
-	}
-}
-
-
-/*
-* キャラについていくNormalAI
-*/
-FollowNormalAI::FollowNormalAI() :
-	NormalAI()
-{
-	m_follow_p = NULL;
-}
-
-Brain* FollowNormalAI::createCopy(std::vector<Character*> characters, const Camera* camera) {
-	FollowNormalAI* res = new FollowNormalAI();
-	for (unsigned int i = 0; i < characters.size(); i++) {
-		if (m_follow_p->getId() == characters[i]->getId()) {
-			res->setFollow(characters[i]);
-			break;
-		}
-	}
-	for (unsigned int i = 0; i < characters.size(); i++) {
-		if (m_target_p->getId() == characters[i]->getId()) {
-			res->setTarget(characters[i]);
-			break;
-		}
-	}
-	setParam(res);
-	return res;
-}
-
-int FollowNormalAI::getFollowId() const {
-	if (m_follow_p == nullptr) { return -1; }
-	return m_follow_p->getId();
-}
-
-void FollowNormalAI::moveOrder(int& right, int& left, int& up, int& down) {
-	// 現在地
-	int x = m_characterAction_p->getCharacter()->getX();
-	int y = m_characterAction_p->getCharacter()->getY();
-
-	// (壁につっかえるなどで)移動できてないから諦める
-	if (m_moveCnt >= GIVE_UP_MOVE_CNT) {
-		m_gx = x;
-		m_gy = y;
-	}
-
-	// 目標地点設定用パラメータ
-	int followX = m_follow_p->getCenterX();
-	bool alreadyGoal = m_gx > x - GX_ERROR && m_gx < x + GX_ERROR;
-	bool alreadyFollow = m_gx < followX + FOLLOW_X_ERROR && m_gx > followX - FOLLOW_X_ERROR;
-
-	// 目標地点設定
-	if ((alreadyGoal && GetRand(MOVE_RAND) == 0) || !alreadyFollow) {
-		if (m_follow_p != NULL) {
-			// followについていく
-			m_gx = m_follow_p->getCenterX() + GetRand(FOLLOW_X_ERROR * 2) - FOLLOW_X_ERROR;
-		}
-		else {
-			// ランダムに設定
-			m_gx = GetRand(200) + 100;
-			if (GetRand(99) < GX_ERROR) { m_gx *= -1; }
-			m_gx += x;
-		}
-	}
-
-	stickOrder(right, left, up, down);
-}
-
-// 追跡対象を決める(AIクラスでオーバライドする。)
-void FollowNormalAI::searchFollow(const Character* character) {
-	// 味方のみ
-	if (character->getId() != m_characterAction_p->getCharacter()->getId() && character->getGroupId() == m_characterAction_p->getCharacter()->getGroupId()) {
-		m_follow_p = character;
-	}
-}
-
-// 追跡対象を変更する必要があるならtrueでアピールする(AIクラスでオーバライドする)。
-bool FollowNormalAI::needSearchFollow() const {
-	if (m_follow_p == NULL || m_follow_p->getHp() == 0) {
-		return true;
-	}
-	return false;
-}
 
 
 /*
@@ -396,6 +21,15 @@ CharacterController::CharacterController(Brain* brain, CharacterAction* characte
 
 	// BrainにActionを注入
 	m_brain->setCharacterAction(m_characterAction);
+
+	// レコーダはデフォルトで使わない
+	m_stickRecorder = NULL;
+	m_jumpRecorder = NULL;
+	m_squatRecorder = NULL;
+	m_slashRecorder = NULL;
+	m_bulletRecorder = NULL;
+
+	m_duplicationFlag = false;
 }
 
 CharacterController::CharacterController() :
@@ -412,7 +46,41 @@ CharacterController::~CharacterController() {
 	if (m_brain != NULL) {
 		delete m_brain;
 	}
+	if (m_stickRecorder != NULL && !m_duplicationFlag) {
+		delete m_stickRecorder;
+		delete m_jumpRecorder;
+		delete m_squatRecorder;
+		delete m_slashRecorder;
+		delete m_bulletRecorder;
+	}
 }
+
+// レコーダを初期化
+void CharacterController::initRecorder() {
+	m_stickRecorder->init();
+	m_jumpRecorder->init();
+	m_squatRecorder->init();
+	m_slashRecorder->init();
+	m_bulletRecorder->init();
+}
+
+// レコードをやめる
+void CharacterController::eraseRecorder() { 
+	if (m_stickRecorder == NULL) { return; }
+	delete m_stickRecorder;
+	delete m_jumpRecorder;
+	delete m_squatRecorder;
+	delete m_slashRecorder;
+	delete m_bulletRecorder;
+	m_stickRecorder = NULL;
+	m_jumpRecorder = NULL;
+	m_squatRecorder = NULL;
+	m_slashRecorder = NULL;
+	m_bulletRecorder = NULL;
+}
+
+// 話しかけたり扉に入ったりするボタンがtrueか
+bool CharacterController::getActionKey() const { return m_brain->actionOrder(); }
 
 void CharacterController::setAction(CharacterAction* action) {
 	delete m_characterAction;
@@ -423,6 +91,27 @@ void CharacterController::setBrain(Brain* brain) {
 	delete m_brain;
 	m_brain = brain;
 	m_brain->setCharacterAction(m_characterAction);
+}
+
+// レコーダのセッタ
+void CharacterController::setStickRecorder(ControllerRecorder* recorder) {
+	m_stickRecorder = recorder;
+}
+void CharacterController::setJumpRecorder(ControllerRecorder* recorder) {
+	m_jumpRecorder = recorder;
+}
+void CharacterController::setSquatRecorder(ControllerRecorder* recorder) {
+	m_squatRecorder = recorder;
+}
+void CharacterController::setSlashRecorder(ControllerRecorder* recorder) {
+	m_slashRecorder = recorder;
+}
+void CharacterController::setBulletRecorder(ControllerRecorder* recorder) {
+	m_bulletRecorder = recorder;
+}
+
+void CharacterController::setTarget(Character* character) { 
+	m_brain->setTarget(character);
 }
 
 // アクションのセッタ
@@ -495,29 +184,104 @@ CharacterController* NormalController::createCopy(std::vector<Character*> charac
 	Brain* brain = m_brain->createCopy(characters, camera);
 	brain->setCharacterAction(action);
 	CharacterController* res = new NormalController(brain, action);
+	// 複製はレコーダをデリートしないためFlagをtrueにする
+	res->setDuplicationFlag(true);
+	res->setStickRecorder(m_stickRecorder);
+	res->setJumpRecorder(m_jumpRecorder);
+	res->setSquatRecorder(m_squatRecorder);
+	res->setSlashRecorder(m_slashRecorder);
+	res->setBulletRecorder(m_bulletRecorder);
 	return res;
 }
 
 void NormalController::control() {
 	// 移動 stickなどの入力状態を更新する
-	int rightStick, leftStick, upStick, downStick;
-	m_brain->moveOrder(rightStick, leftStick, upStick, downStick);
+	int rightStick = 0, leftStick = 0, upStick = 0, downStick = 0;
+	if (m_stickRecorder != NULL) {
+		if (m_stickRecorder->existRecord()) {
+			int input = m_stickRecorder->checkInput();
+			if (((input >> 0) & 1) == 1) { 
+				rightStick = 1;
+			}
+			if (((input >> 1) & 1) == 1) { 
+				leftStick = 1;
+			}
+			if (((input >> 2) & 1) == 1) { upStick = 1; }
+			if (((input >> 3) & 1) == 1) { downStick = 1; }
+		}
+		else {
+			m_brain->moveOrder(rightStick, leftStick, upStick, downStick);
+			int bit = 0;
+			if (rightStick > 0) { bit |= (1 << 0); }
+			if (leftStick > 0) { bit |= (1 << 1); }
+			if (upStick > 0) { bit |= (1 << 2); }
+			if (downStick > 0) { bit |= (1 << 3); }
+			m_stickRecorder->writeRecord(bit);
+		}
+		m_stickRecorder->addTime();
+	}
+	else {
+		m_brain->moveOrder(rightStick, leftStick, upStick, downStick);
+	}
 
 	// stickに応じて移動
 	m_characterAction->move(rightStick, leftStick, upStick, downStick);
 
 	// ジャンプ
-	int jump = m_brain->jumpOrder();
+	int jump = 0;
+	if (m_jumpRecorder != NULL) {
+		if (m_jumpRecorder->existRecord()) {
+			jump = m_jumpRecorder->checkInput();
+		}
+		else {
+			jump = m_brain->jumpOrder();
+			if (jump == 1) {
+				jump;
+			}
+			m_jumpRecorder->writeRecord(jump);
+		}
+		m_jumpRecorder->addTime();
+	}
+	else {
+		jump = m_brain->jumpOrder();
+	}
 	m_characterAction->jump(jump);
 
 	// しゃがみ
-	int squat = m_brain->squatOrder();
+	int squat = 0;
+	if (m_squatRecorder != NULL) {
+		if (m_squatRecorder->existRecord()) {
+			squat = m_squatRecorder->checkInput();
+		}
+		else {
+			squat = m_brain->squatOrder();
+			m_squatRecorder->writeRecord(squat);
+		}
+		m_squatRecorder->addTime();
+	}
+	else {
+		squat = m_brain->squatOrder();
+	}
 	m_characterAction->setSquat(squat);
 }
 
 Object* NormalController::bulletAttack() {
+	int order = 0;
+	if (m_bulletRecorder != NULL) {
+		if (m_bulletRecorder->existRecord()) {
+			order = m_bulletRecorder->checkInput();
+		}
+		else {
+			order = m_brain->bulletOrder();
+			m_bulletRecorder->writeRecord(order);
+		}
+		m_bulletRecorder->addTime();
+	}
+	else {
+		order = m_brain->bulletOrder();
+	}
 	// 遠距離攻撃の命令がされているなら
-	if (m_brain->bulletOrder() > 0) {
+	if (order > 0) {
 		// 攻撃目標
 		int targetX, targetY;
 		m_brain->bulletTargetPoint(targetX, targetY);
@@ -528,8 +292,22 @@ Object* NormalController::bulletAttack() {
 }
 
 Object* NormalController::slashAttack() {
+	int order = 0;
+	if (m_slashRecorder != NULL) {
+		if (m_slashRecorder->existRecord()) {
+			order = m_slashRecorder->checkInput();
+		}
+		else {
+			order = m_brain->slashOrder();
+			m_slashRecorder->writeRecord(order);
+		}
+		m_slashRecorder->addTime();
+	}
+	else {
+		order = m_brain->slashOrder();
+	}
 	// 近距離攻撃の命令がされたか、した後で今が攻撃タイミングなら
-	if (m_brain->slashOrder() == 1 || m_characterAction->getSlashCnt() > 0) {
+	if (order == 1 || m_characterAction->getSlashCnt() > 0) {
 		// 攻撃目標
 		int targetX, targetY;
 		m_brain->bulletTargetPoint(targetX, targetY);
